sql: recalculate statement ID on collision
Before this patch, there was a possibility that multiple different
queries could share the same prepared statement ID. This could lead to
lost prepared statements. This patch fixes this issue.

Closes #12537

NO_DOC=bugfix

(cherry picked from commit 42706a28c768bbc04ebf9da0d9769a49657dad20)

+++ src/box/execute.h
@@ -120,7 +120,7 @@ sql_stmt_busy(const struct Vdbe *stmt);
  * @param port Port to store request response.
  */
 int
-sql_prepare(const char *sql, int len, struct port *port);
+sql_prepare(const char *sql, size_t len, struct port *port);
 
 /**
  * Process an SQL request received over IPROTO.
